requirements.txt
asgiref==3.8.1
Django==4.2.21
django-filter==23.5
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
inflection==0.5.1
packaging==26.0
psycopg2-binary==2.9.9
PyJWT==2.10.1
PyMySQL==1.1.1
python-dotenv==1.0.0
pytz==2026.1.post1
pywin32==306
PyYAML==6.0.3
setuptools==75.4.0
sqlparse==0.5.3
supervisor==4.2.5
supervisor-win==4.7.0
tzdata==2024.2
uritemplate==4.2.0
